Completely rewrite originalElements into StatefulDom.
StatefulDom (instanced as domWrapper) keeps the state of the dom as
it should be and initially was.  This lets it revert any changes made
by the plugin, as well as reenforce the plugin's changes if the dom
was modified by third party code.  It currenty does this reenforcement
on window resize, but in the future it may implement a dom watcher to
automatically revert changes.